PC is uint64 since that's how big mem is

+++ inc/MachineState.hpp
@@ -11,7 +11,7 @@ public:
   std::vector<Register> registerFile;
   std::map<uint64_t, Register> memory;
 
-  uint32_t pc;
+  uint64_t pc;
   bool running;
 
   MachineState(uint64_t starting_addr);
